Reduce memory allocations by native libraries.
NativeCipher and NativeHash's backing native code don't store any state (they always use contexts), so it is safe to share instances. (In theory, NativeZlib can use a ThreadLocal, but Voxelwind already does this.)

This is more apparent with NativeHash as an instance is created every time a packet is sent.